[restructure] temporarily moved demo files into one folder to not be blocked by import problems
