--- demos/pycram_virtual_building_demos/README.md
@@ -1,1 +1,0 @@
-
