+++ demos/pycram_virtual_building_demos/hsrb_navigate_demo.py
@@ -1,9 +1,6 @@
-from demos.pycram_virtual_building_demos.setup.launch_robot import launch_tiago
-from demos.pycram_virtual_building_demos.setup.utils import display_loading_gif
+from demos.pycram_virtual_building_demos.setup_launch_robot import launch_hsrb
 from pycram.datastructures.enums import ObjectType, WorldMode
-from pycram.datastructures.pose import Pose
 from pycram.designators.action_designator import *
-from pycram.designators.location_designator import *
 from pycram.designators.object_designator import *
 from pycram.object_descriptors.urdf import ObjectDescription
 from pycram.process_module import simulated_robot
@@ -17,27 +14,28 @@ def navigate_main():
 
     world = BulletWorld(WorldMode.DIRECT)
     viz_marker = VizMarkerPublisher()
-    robot = Object("tiago_dual", ObjectType.ROBOT, f"tiago_dual{extension}", pose=Pose([1, 2, 0]))
+    robot = Object("hsrb", ObjectType.ROBOT, f"hsrb{extension}", pose=Pose([1, 2, 0]))
     apartment = Object("apartment", ObjectType.ENVIRONMENT, f"apartment{extension}")
 
-    robot_desig = BelieveObject(names=["tiago_dual"])
+    robot_desig = BelieveObject(names=["hsrb"])
     apartment_desig = BelieveObject(names=["apartment"])
 
     navigate_pose = Pose([2.7, 2.15, 1])
 
     with simulated_robot:
-        ParkArmsAction([Arms.BOTH]).resolve().perform()
-
         MoveTorsoAction([0.25]).resolve().perform()
 
         NavigateAction([navigate_pose]).resolve().perform()
 
 
 def navigate_demo():
-    launch_tiago()
+    launch_hsrb()
+
+    rospy.sleep(5)
 
     navigate_main()
 
 
 if __name__ == "__main__":
-    navigate_demo()
+    navigate_main()
+
