+++ demos/pycram_virtual_building_demos/pr2_navigate_demo.py
@@ -1,4 +1,3 @@
-from demos.pycram_virtual_building_demos.setup.launch_robot import launch_pr2
 from pycram.datastructures.enums import ObjectType, WorldMode
 from pycram.designators.action_designator import *
 from pycram.designators.location_designator import *
@@ -32,10 +31,6 @@ def navigate_main():
 
 
 def navigate_demo():
-    launch_pr2()
-
-    rospy.sleep(3)
-
     navigate_main()
