+++ demos/pycram_virtual_building_demos/setup_launch_robot.py
