+++ demos/pycram_virtual_building_demos/setup_utils.py
@@ -7,6 +7,7 @@ from ipywidgets import HTML
 def display_loading_gif():
     import base64
     setup_path = os.path.dirname(os.path.realpath(__file__))
-    gif_path = setup_path + '/imgs/loading.gif'
+    #gif_path = setup_path + '/imgs/loading.gif'
+    gif_path = setup_path + '/setup/imgs/loading.gif'
     b64 = base64.b64encode(open(gif_path, 'rb').read()).decode('ascii')
     display(HTML(f'<img src="data:image/gif;base64,{b64}" />'))
